feat(db): ship the kick/db typegen on dbCliPlugin
Move kickDbTypegen into @forinda/kickjs-db/cli and add it to
dbCliPlugin.typegens, so mounting the plugin wires both the `kick db`
commands and `.kickjs/types/kick__db.d.ts` generation. The typegen is
typed against the @forinda/kickjs-cli-kit CliTypegen contract (reads only
ctx.cwd), so it needs no kickjs-cli internals.

kickjs-cli's builtin/db.ts becomes a re-export shim (back-compat) and the
typegen is no longer auto-registered — it comes with the opt-in plugin.

Verified: `kick typegen` in the SQLite example emits kick__db.d.ts via the
mounted plugin. db 413 tests, cli 381 tests.
